2021/12/10 ChatActivity 구조 변경 --- 김민정
<개발내용>
1. 기존 postActivity에서 roomId를 찾아서 intent로 넘겨주는 방식에서 ChatActivity 내부에서 직접 찾는 방식으로 변경. 단 대화하는 상대방의 id 정보는 intent로 넘겨받아야 함.
