+++ android/app/google-services.json
@@ -1,0 +1,64 @@
+{
+  "project_info": {
+    "project_number": "81283441548",
+    "firebase_url": "https://shareeats-5654f-default-rtdb.firebaseio.com",
+    "project_id": "shareeats-5654f",
+    "storage_bucket": "shareeats-5654f.appspot.com"
+  },
+  "client": [
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:81283441548:android:73f3b782920e5aaa65aca9",
+        "android_client_info": {
+          "package_name": "com.sharewanted.shareeats"
+        }
+      },
+      "oauth_client": [
+        {
+          "client_id": "81283441548-4nt3djqv9hk6qidfehla2mkm2u90bv81.apps.googleusercontent.com",
+          "client_type": 1,
+          "android_info": {
+            "package_name": "com.sharewanted.shareeats",
+            "certificate_hash": "bbcb35b5c171647763f6eeca148ade92aa91e783"
+          }
+        },
+        {
+          "client_id": "81283441548-j6ibeomllpks2map9ktg43p7dkl7fome.apps.googleusercontent.com",
+          "client_type": 1,
+          "android_info": {
+            "package_name": "com.sharewanted.shareeats",
+            "certificate_hash": "7e6794736f16d0d84f4e655f885c8d9254e90a55"
+          }
+        },
+        {
+          "client_id": "81283441548-ji9b10a8ena0o5chr5fbp3rlsk9f5ca8.apps.googleusercontent.com",
+          "client_type": 1,
+          "android_info": {
+            "package_name": "com.sharewanted.shareeats",
+            "certificate_hash": "1c592235a92c55539222e4fd7669739eaccdfc12"
+          }
+        },
+        {
+          "client_id": "81283441548-76rvbp0pk38f33s595gef858abb96c3d.apps.googleusercontent.com",
+          "client_type": 3
+        }
+      ],
+      "api_key": [
+        {
+          "current_key": "AIzaSyDuEREGQof0xS0Z8NA_GhSgcDlXMWRyIEw"
+        }
+      ],
+      "services": {
+        "appinvite_service": {
+          "other_platform_oauth_client": [
+            {
+              "client_id": "81283441548-76rvbp0pk38f33s595gef858abb96c3d.apps.googleusercontent.com",
+              "client_type": 3
+            }
+          ]
+        }
+      }
+    }
+  ],
+  "configuration_version": "1"
+}
